Rename certificates to certficiate Rename applications to application
