+++ application/datastore.go
@@ -1,4 +1,4 @@
-package applications
+package application
 
 import (
 	"fmt"
